Rename `Enum` namespace to `Enumeration`
`Enum` is a soft-reserved keyword by PHP
Closes #3
